change component names
